feat: 그룹원 엔티티 추가 (#464) - flyway 버전 추가
